// FrontEnd/src/app/models/Filtros/FiltroPadrao.ts
export class FiltroPadrao {

    constructor() {
        this.Target = new Array<PadraoComboFiltro>();
        this.Regiao = new Array<PadraoComboFiltro>();
        this.Demografico = new Array<PadraoComboFiltro>();
        this.Onda = new Array<PadraoComboFiltro>();
        this.Marca = new Array<PadraoComboFiltro>();
    }

    Target: Array<PadraoComboFiltro>;
    Regiao: Array<PadraoComboFiltro>;
    Demografico: Array<PadraoComboFiltro>;
    Onda: Array<PadraoComboFiltro>;
    Marca: Array<PadraoComboFiltro>;
    CodUser: number;
    CodIdioma: number;
    Sequencia: number;
    Atributo: number;
    TituloGrafico: string;
    ParamTipo:number;
    ParamOndaAtual :number;
    ParamSTB:number;
    ParamBIA:number;
    ParamDenominators:number;
}